Remove Node 6 and RethinkDB support
