+++ .travis.yml
@@ -4,5 +4,3 @@ node_js:
   - 8
   - 10
 install: npm install
-addons:
-  rethinkdb: 2.3
