+++ generators/hook/index.js
@@ -162,12 +162,12 @@ module.exports = class HookGenerator extends Generator {
     }
 
     this.fs.copyTpl(
-      this.templatePath(this.hasAsync ? 'hook-async.js' : 'hook.js'),
+      this.templatePath('hook.js'),
       mainFile, context
     );
 
     this.fs.copyTpl(
-      this.templatePath(this.hasAsync ? `test-async.${tester}.js` : `test.${tester}.js`),
+      this.templatePath(`test.${tester}.js`),
       this.destinationPath(this.testDirectory, 'hooks', `${context.kebabName}.test.js`),
       context
     );
