--- generators/hook/templates/hook-async.js
@@ -1,9 +1,0 @@
-// Use this hook to manipulate incoming or outgoing data.
-// For more information on hooks see: http://docs.feathersjs.com/api/hooks.html
-
-// eslint-disable-next-line no-unused-vars
-module.exports = function (options = {}) {
-  return async context => {
-    return context;
-  };
-};
