+++ generators/hook/templates/hook.js
@@ -1,8 +1,9 @@
 // Use this hook to manipulate incoming or outgoing data.
 // For more information on hooks see: http://docs.feathersjs.com/api/hooks.html
 
-module.exports = function (options = {}) { // eslint-disable-line no-unused-vars
-  return context => {
-    return Promise.resolve(context);
+// eslint-disable-next-line no-unused-vars
+module.exports = function (options = {}) {
+  return async context => {
+    return context;
   };
 };
