+++ generators/hook/templates/test.mocha.js
@@ -9,8 +9,8 @@ describe('\'<%= name %>\' hook', () => {
     app = feathers();
 
     app.use('/dummy', {
-      get(id) {
-        return Promise.resolve({ id });
+      async get(id) {
+        return { id };
       }
     });
 
@@ -19,9 +19,9 @@ describe('\'<%= name %>\' hook', () => {
     });
   });
 
-  it('runs the hook', () => {
-    return app.service('dummy').get('test').then(result => {
-      assert.deepEqual(result, { id: 'test' });
-    });
+  it('runs the hook', async () => {
+    const result = await app.service('dummy').get('test');
+    
+    assert.deepEqual(result, { id: 'test' });
   });
 });
