--- generators/service/templates/class-async.js
@@ -1,42 +1,0 @@
-/* eslint-disable no-unused-vars */
-class Service {
-  constructor (options) {
-    this.options = options || {};
-  }
-
-  async find (params) {
-    return [];
-  }
-
-  async get (id, params) {
-    return {
-      id, text: `A new message with ID: ${id}!`
-    };
-  }
-
-  async create (data, params) {
-    if (Array.isArray(data)) {
-      return Promise.all(data.map(current => this.create(current, params)));
-    }
-
-    return data;
-  }
-
-  async update (id, data, params) {
-    return data;
-  }
-
-  async patch (id, data, params) {
-    return data;
-  }
-
-  async remove (id, params) {
-    return { id };
-  }
-}
-
-module.exports = function (options) {
-  return new Service(options);
-};
-
-module.exports.Service = Service;
