+++ lib/generator.js
@@ -35,14 +35,6 @@ module.exports = class BaseGenerator extends Generator {
     }
   }
 
-  get hasAsync() {
-    const asyncNode = '>= 8.0.0';
-    const pkgEngineAsync = !(this.pkg.engines && this.pkg.engines.node) ||
-      semver.intersects(this.pkg.engines.node, asyncNode);
-
-    return semver.satisfies(process.version, asyncNode) && pkgEngineAsync;
-  }
-
   get libDirectory() {
     return this.pkg.directories && this.pkg.directories.lib;
   }
