+++ package.json
@@ -79,7 +79,6 @@
     "feathers-mongoose": "^7.0.0",
     "feathers-nedb": "^4.0.0",
     "feathers-objection": "^3.0.0",
-    "feathers-rethinkdb": "^0.5.3",
     "feathers-sequelize": "^5.0.0",
     "helmet": "^3.15.0",
     "jshint": "^2.9.6",
